moving assertion related primitive types to assertion module
