+++ primitives/core/src/assertion/platform_user.rs
@@ -18,7 +18,7 @@ use parity_scale_codec::{Decode, Encode, MaxEncodedLen};
 use scale_info::TypeInfo;
 use sp_std::vec::Vec;
 
-use crate::{all_evm_web3networks, Web3Network};
+use crate::assertion::network::{all_evm_web3networks, Web3Network};
 
 #[derive(Encode, Decode, Clone, Debug, PartialEq, Eq, MaxEncodedLen, TypeInfo)]
 pub enum PlatformUserType {
